split off wand module
